pyproject.toml
[project]
name = "lab10"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "mysql-connector-python",
    "httpx",
    "python-dotenv>=1.2.2",
    "jinja2",
    "python-multipart",
    "asyncio>=4.0.0",
]
